Add subordinates with docs to the declaration map
The only place in the code where we want the subordinates for a declaration is
right after having looked up the declaration in the map.

And since we include subordinates in the map, we might as well take the
opportunity to store those subordinates that belong to a particular declaration
together with that declaration.

We also store the documentation for each subordinate.